+++ src/main/java/com/icestudyroom_email/domain/infrastructure/email/template/RankingEmailTemplateResolver.java
@@ -1,7 +1,7 @@
-package com.icestudyroom_email.domain.email.infrastructure.template;
+package com.icestudyroom_email.domain.infrastructure.email.template;
 
-import com.icestudyroom_email.domain.email.infrastructure.gmail.dto.EmailRequest;
-import com.icestudyroom_email.domain.rankingContract.RankingEmailEvent;
+import com.icestudyroom_email.domain.infrastructure.email.gmail.EmailRequest;
+import com.icestudyroom_email.domain.contract.ranking.RankingEmailEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
